Игрок переработан
объект игрока теперь является членом IPattern Player, а не отдельно обрабатываемой сущностью. инициализатор и апдейтер игрока теперь часть класса Player

+++ VideoGame/Construct/Processing/GameCamera.cs
@@ -20,7 +20,7 @@ namespace VideoGame
 
         public Vector2 ApplyParalax(Vector2 position, float dx, float dy)
         {
-            return position - new Vector2(LeftTopCorner.X * dx, LeftTopCorner.Y * dy);
+            return position - new Vector2((int)(LeftTopCorner.X * dx), (int)(LeftTopCorner.Y * dy));
         }
         public GameObject TargetObject { get; private set; }
         public Rectangle InnerBorders { get; private set; }
